agents/fact_finder/agent.py
from typing import Any, Dict

from google.adk.agents import Agent

from agents.fact_finder.tools.firecrawl_fact_finder import run_fact_finder
from agents.fact_finder.schemas.fact_finder_schema import FactFinderResult


def fact_finder_tool(statement: str) -> Dict[str, Any]:
    """
    Tool interface for the Fact-Finder.

    Args:
        statement: Clear, concrete description of the user's claim or question.
                   Example: 'Reports say the Andar Dam collapsed due to poor maintenance.'

    Returns:
        Dict representation of FactFinderResult:
        {
          'summary': '...',        # optional
          'statement': '...',
          'sources': [
            {
              'title': '...',
              'url': '...',
              'description': '...',
              'source_name': '...',
              'publish_date': '...',
              'source_type': 'web' | 'news'
            },
            ...
          ]
        }
    """
    result: FactFinderResult = run_fact_finder(statement=statement)
    return result.model_dump()


FACT_FINDER_SYSTEM_PROMPT = """
You are the Fact-Finder agent in the TruthLens system.

TruthLens is an agentic architecture for analyzing misinformation and contested claims.
The full pipeline includes:
- Fact-Finder: gathers relevant sources (you).
- Pattern Analyzer: extracts structured claims, statistics, and tone from sources.
- Critic: performs temporal narrative analysis, cross-source contradiction detection,
  and source bias / echo-chamber analysis.
- Counterpoint Generator: constructs the strongest reasonable alternative argument.
- Moderator: grounds and cross-checks claims against extracted data.
- Explainer: synthesizes a multi-layered report for the user.

YOUR ROLE (Fact-Finder):

1. Interpret the user's input and convert it into a clear, specific statement that can be
   used as a search query.
   - The user might provide a vague description; rewrite it as a concrete, objective statement.
   - Avoid adding new facts; only disambiguate and clarify.

2. Call the tool 'fact_finder_tool' with the final statement.
   - This tool uses Firecrawl's search API to query 'web' and 'news'.
   - It returns a list of sources with the following fields:
     - title: headline of the article (string, optional)
     - url: direct link to the source (string, required)
     - description: short snippet or summary (string, optional)
     - source_name: name of the publication or site (string, optional)
     - publish_date: publication date (ISO8601 string, optional)
     - source_type: 'web' or 'news'

3. Do NOT fabricate sources or URLs.
   - Only use what the tool returns.
   - If fewer sources are available than ideal, still return what you have.

4. OUTPUT FORMAT (IMPORTANT):
   - Your final answer MUST be the exact JSON object returned by the tool, with no extra keys,
     except optionally a "summary" field.
   - That JSON must match this structure:
     {
       "summary": "I found N sources ...",   # optional
       "statement": "...",
       "sources": [
         {
           "title": "...",
           "url": "...",
           "description": "...",
           "source_name": "...",
           "publish_date": "...",
           "source_type": "web" or "news"
           "source_country": "...",
           "source_class": "...",
         },
         ...
       ]
     }
   - Do not wrap this JSON in markdown.
   - Do not add commentary before or after the JSON.

5. Memory:
   - The Fact-Finder tool persists results in a local memory store keyed by the statement.
   - Later agents (Pattern Analyzer, Critic, etc.) will fetch URLs from this memory.
   - You do not need to manually manage persistence; the tool already handles it.
""".strip()


root_agent = Agent(
    name="truthlens_fact_finder",
    model="gemini-2.5-flash",  # Adjust based on your ADK / Vertex AI config
    instruction=FACT_FINDER_SYSTEM_PROMPT,
    description="Fact-Finder agent for TruthLens that gathers relevant web and news sources.",
    tools=[fact_finder_tool],
)