requirements.txt
google-adk
requests
pydantic
python-dotenv
